fixed the issue with backspace, by factoring out Easy(1|0)
  * Backspace previously would not re-enabled. This is fixed
  * We previously had two functions we could call that would call one
function which took a binary argument. This binary argument would have
logic that would essentially infer which one of the two functions call
it and impliment the body of what should have been those functions. That
was nasty and silly and hard to follow. I've removed Easy(1) and Easy(0)
entirely and made the code more consistent in the process.